backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
langchain==0.3.27
langchain-community==0.3.27
langchain-google-genai==2.1.9
langchain-huggingface==0.3.1
faiss-cpu==1.11.0.post1
sentence-transformers==5.0.0
google-generativeai==0.8.5
python-dotenv==1.1.1